- Move Shop and Login to /views
