+++ client/src/renderer/ui/views/Login.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../../../context/AuthProvider";
+import { useAuth } from "../../context/AuthProvider";
 
 const LoginPage = () => {
   const [username, setUsername] = useState("");
